faculty update completed with basic view

+++ SmartStudent/src/main/java/com/manage/faculty/service/FacultyAddressService.java
@@ -3,7 +3,7 @@ package com.manage.faculty.service;
 import com.manage.faculty.entities.FacultyAddress;
 
 public interface FacultyAddressService {
-    FacultyAddress getAddressById(int id);
+    FacultyAddress getAddressById(Long id);
     void saveAddress(FacultyAddress address);
-    void deleteAddress(int id);
+    void deleteAddress(Long id);
 }
